pip-requirements.txt
pytest
PyYAML
termcolor
coverage
